Module amazon_sp_catalog_items::apis::catalog_api

Functions

get_catalog_item

Effective September 30, 2022, the getCatalogItem operation will no longer be available in the Selling Partner API for Catalog Items v0. This operation is available in the latest version of the Selling Partner API for Catalog Items v2022-04-01. Integrations that rely on this operation should migrate to the latest version to avoid service disruption. Note: The listCatalogCategories operation is not being deprecated and you can continue to make calls to it.

list_catalog_categories

Returns the parent categories to which an item belongs, based on the specified ASIN or SellerSKU. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 2 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

list_catalog_items

Effective September 30, 2022, the listCatalogItems operation will no longer be available in the Selling Partner API for Catalog Items v0. As an alternative, searchCatalogItems is available in the latest version of the Selling Partner API for Catalog Items v2022-04-01. Integrations that rely on the listCatalogItems operation should migrate to the searchCatalogItemsoperation to avoid service disruption. Note: The listCatalogCategories operation is not being deprecated and you can continue to make calls to it.