Function amazon_sp_catalog_items::apis::catalog_api::list_catalog_items

pub async fn list_catalog_items(
    configuration: &Configuration,
    marketplace_id: &str,
    query: Option<&str>,
    query_context_id: Option<&str>,
    seller_sku: Option<&str>,
    UPC: Option<&str>,
    EAN: Option<&str>,
    ISBN: Option<&str>,
    JAN: Option<&str>
) -> Result<ListCatalogItemsResponse, Error>

Effective September 30, 2022, the listCatalogItems operation will no longer be available in the Selling Partner API for Catalog Items v0. As an alternative, searchCatalogItems is available in the latest version of the Selling Partner API for Catalog Items v2022-04-01. Integrations that rely on the listCatalogItems operation should migrate to the searchCatalogItemsoperation to avoid service disruption. Note: The listCatalogCategories operation is not being deprecated and you can continue to make calls to it.