Struct amazon_sp_catalog_items::models::attribute_set_list_type::AttributeSetListType

pub struct AttributeSetListType {

    pub actor: Option<Vec<String>>,
    pub artist: Option<Vec<String>>,
    pub aspect_ratio: Option<String>,
    pub audience_rating: Option<String>,
    pub author: Option<Vec<String>>,
    pub back_finding: Option<String>,
    pub band_material_type: Option<String>,
    pub binding: Option<String>,
    pub bluray_region: Option<String>,
    pub brand: Option<String>,
    pub cero_age_rating: Option<String>,
    pub chain_type: Option<String>,
    pub clasp_type: Option<String>,
    pub color: Option<String>,
    pub cpu_manufacturer: Option<String>,
    pub cpu_speed: Option<Box<DecimalWithUnits>>,
    pub cpu_type: Option<String>,
    pub creator: Option<Vec<CreatorType>>,
    pub department: Option<String>,
    pub director: Option<Vec<String>>,
    pub display_size: Option<Box<DecimalWithUnits>>,
    pub edition: Option<String>,
    pub episode_sequence: Option<String>,
    pub esrb_age_rating: Option<String>,
    pub feature: Option<Vec<String>>,
    pub flavor: Option<String>,
    pub format: Option<Vec<String>>,
    pub gem_type: Option<Vec<String>>,
    pub genre: Option<String>,
    pub golf_club_flex: Option<String>,
    pub golf_club_loft: Option<Box<DecimalWithUnits>>,
    pub hand_orientation: Option<String>,
    pub hard_disk_interface: Option<String>,
    pub hard_disk_size: Option<Box<DecimalWithUnits>>,
    pub hardware_platform: Option<String>,
    pub hazardous_material_type: Option<String>,
    pub item_dimensions: Option<Box<DimensionType>>,
    pub is_adult_product: Option<bool>,
    pub is_autographed: Option<bool>,
    pub is_eligible_for_trade_in: Option<bool>,
    pub is_memorabilia: Option<bool>,
    pub issues_per_year: Option<String>,
    pub item_part_number: Option<String>,
    pub label: Option<String>,
    pub languages: Option<Vec<LanguageType>>,
    pub legal_disclaimer: Option<String>,
    pub list_price: Option<Box<Price>>,
    pub manufacturer: Option<String>,
    pub manufacturer_maximum_age: Option<Box<DecimalWithUnits>>,
    pub manufacturer_minimum_age: Option<Box<DecimalWithUnits>>,
    pub manufacturer_parts_warranty_description: Option<String>,
    pub material_type: Option<Vec<String>>,
    pub maximum_resolution: Option<Box<DecimalWithUnits>>,
    pub media_type: Option<Vec<String>>,
    pub metal_stamp: Option<String>,
    pub metal_type: Option<String>,
    pub model: Option<String>,
    pub number_of_discs: Option<i32>,
    pub number_of_issues: Option<i32>,
    pub number_of_items: Option<i32>,
    pub number_of_pages: Option<i32>,
    pub number_of_tracks: Option<i32>,
    pub operating_system: Option<Vec<String>>,
    pub optical_zoom: Option<Box<DecimalWithUnits>>,
    pub package_dimensions: Option<Box<DimensionType>>,
    pub package_quantity: Option<i32>,
    pub part_number: Option<String>,
    pub pegi_rating: Option<String>,
    pub platform: Option<Vec<String>>,
    pub processor_count: Option<i32>,
    pub product_group: Option<String>,
    pub product_type_name: Option<String>,
    pub product_type_subcategory: Option<String>,
    pub publication_date: Option<String>,
    pub publisher: Option<String>,
    pub region_code: Option<String>,
    pub release_date: Option<String>,
    pub ring_size: Option<String>,
    pub running_time: Option<Box<DecimalWithUnits>>,
    pub shaft_material: Option<String>,
    pub scent: Option<String>,
    pub season_sequence: Option<String>,
    pub seikodo_product_code: Option<String>,
    pub size: Option<String>,
    pub size_per_pearl: Option<String>,
    pub small_image: Option<Box<Image>>,
    pub studio: Option<String>,
    pub subscription_length: Option<Box<DecimalWithUnits>>,
    pub system_memory_size: Option<Box<DecimalWithUnits>>,
    pub system_memory_type: Option<String>,
    pub theatrical_release_date: Option<String>,
    pub title: Option<String>,
    pub total_diamond_weight: Option<Box<DecimalWithUnits>>,
    pub total_gem_weight: Option<Box<DecimalWithUnits>>,
    pub warranty: Option<String>,
    pub weee_tax_value: Option<Box<Price>>,

}

AttributeSetListType : The attributes of the item.

Fields
actor: Option<Vec<String>>

The actor attributes of the item.

artist: Option<Vec<String>>

The artist attributes of the item.

aspect_ratio: Option<String>

The aspect ratio attribute of the item.

audience_rating: Option<String>

The audience rating attribute of the item.

author: Option<Vec<String>>

The author attributes of the item.

back_finding: Option<String>

The back finding attribute of the item.

band_material_type: Option<String>

The band material type attribute of the item.

binding: Option<String>

The binding attribute of the item.

bluray_region: Option<String>

The Bluray region attribute of the item.

brand: Option<String>

The brand attribute of the item.

cero_age_rating: Option<String>

The CERO age rating attribute of the item.

chain_type: Option<String>

The chain type attribute of the item.

clasp_type: Option<String>

The clasp type attribute of the item.

color: Option<String>

The color attribute of the item.

cpu_manufacturer: Option<String>

The CPU manufacturer attribute of the item.

cpu_speed: Option<Box<DecimalWithUnits>>cpu_type: Option<String>

The CPU type attribute of the item.

creator: Option<Vec<CreatorType>>

The creator attributes of the item.

department: Option<String>

The department attribute of the item.

director: Option<Vec<String>>

The director attributes of the item.

display_size: Option<Box<DecimalWithUnits>>edition: Option<String>

The edition attribute of the item.

episode_sequence: Option<String>

The episode sequence attribute of the item.

esrb_age_rating: Option<String>

The ESRB age rating attribute of the item.

feature: Option<Vec<String>>

The feature attributes of the item

flavor: Option<String>

The flavor attribute of the item.

format: Option<Vec<String>>

The format attributes of the item.

gem_type: Option<Vec<String>>

The gem type attributes of the item.

genre: Option<String>

The genre attribute of the item.

golf_club_flex: Option<String>

The golf club flex attribute of the item.

golf_club_loft: Option<Box<DecimalWithUnits>>hand_orientation: Option<String>

The hand orientation attribute of the item.

hard_disk_interface: Option<String>

The hard disk interface attribute of the item.

hard_disk_size: Option<Box<DecimalWithUnits>>hardware_platform: Option<String>

The hardware platform attribute of the item.

hazardous_material_type: Option<String>

The hazardous material type attribute of the item.

item_dimensions: Option<Box<DimensionType>>is_adult_product: Option<bool>

The adult product attribute of the item.

is_autographed: Option<bool>

The autographed attribute of the item.

is_eligible_for_trade_in: Option<bool>

The is eligible for trade in attribute of the item.

is_memorabilia: Option<bool>

The is memorabilia attribute of the item.

issues_per_year: Option<String>

The issues per year attribute of the item.

item_part_number: Option<String>

The item part number attribute of the item.

label: Option<String>

The label attribute of the item.

languages: Option<Vec<LanguageType>>

The languages attribute of the item.

legal_disclaimer: Option<String>

The legal disclaimer attribute of the item.

list_price: Option<Box<Price>>manufacturer: Option<String>

The manufacturer attribute of the item.

manufacturer_maximum_age: Option<Box<DecimalWithUnits>>manufacturer_minimum_age: Option<Box<DecimalWithUnits>>manufacturer_parts_warranty_description: Option<String>

The manufacturer parts warranty description attribute of the item.

material_type: Option<Vec<String>>

The material type attributes of the item.

maximum_resolution: Option<Box<DecimalWithUnits>>media_type: Option<Vec<String>>

The media type attributes of the item.

metal_stamp: Option<String>

The metal stamp attribute of the item.

metal_type: Option<String>

The metal type attribute of the item.

model: Option<String>

The model attribute of the item.

number_of_discs: Option<i32>

The number of discs attribute of the item.

number_of_issues: Option<i32>

The number of issues attribute of the item.

number_of_items: Option<i32>

The number of items attribute of the item.

number_of_pages: Option<i32>

The number of pages attribute of the item.

number_of_tracks: Option<i32>

The number of tracks attribute of the item.

operating_system: Option<Vec<String>>

The operating system attributes of the item.

optical_zoom: Option<Box<DecimalWithUnits>>package_dimensions: Option<Box<DimensionType>>package_quantity: Option<i32>

The package quantity attribute of the item.

part_number: Option<String>

The part number attribute of the item.

pegi_rating: Option<String>

The PEGI rating attribute of the item.

platform: Option<Vec<String>>

The platform attributes of the item.

processor_count: Option<i32>

The processor count attribute of the item.

product_group: Option<String>

The product group attribute of the item.

product_type_name: Option<String>

The product type name attribute of the item.

product_type_subcategory: Option<String>

The product type subcategory attribute of the item.

publication_date: Option<String>

The publication date attribute of the item.

publisher: Option<String>

The publisher attribute of the item.

region_code: Option<String>

The region code attribute of the item.

release_date: Option<String>

The release date attribute of the item.

ring_size: Option<String>

The ring size attribute of the item.

running_time: Option<Box<DecimalWithUnits>>shaft_material: Option<String>

The shaft material attribute of the item.

scent: Option<String>

The scent attribute of the item.

season_sequence: Option<String>

The season sequence attribute of the item.

seikodo_product_code: Option<String>

The Seikodo product code attribute of the item.

size: Option<String>

The size attribute of the item.

size_per_pearl: Option<String>

The size per pearl attribute of the item.

small_image: Option<Box<Image>>studio: Option<String>

The studio attribute of the item.

subscription_length: Option<Box<DecimalWithUnits>>system_memory_size: Option<Box<DecimalWithUnits>>system_memory_type: Option<String>

The system memory type attribute of the item.

theatrical_release_date: Option<String>

The theatrical release date attribute of the item.

title: Option<String>

The title attribute of the item.

total_diamond_weight: Option<Box<DecimalWithUnits>>total_gem_weight: Option<Box<DecimalWithUnits>>warranty: Option<String>

The warranty attribute of the item.

weee_tax_value: Option<Box<Price>>
Implementations

impl AttributeSetListType

pub fn new() -> AttributeSetListType

The attributes of the item.