Function amazon_sp_catalog_items::apis::catalog_api::list_catalog_items

pub async fn list_catalog_items(
    configuration: &Configuration, 
    marketplace_id: &str, 
    query: Option<&str>, 
    query_context_id: Option<&str>, 
    seller_sku: Option<&str>, 
    UPC: Option<&str>, 
    EAN: Option<&str>, 
    ISBN: Option<&str>, 
    JAN: Option<&str>
) -> Result<ListCatalogItemsResponse, Error<ListCatalogItemsError>>

Returns a list of items and their attributes, based on a search query or item identifiers that you specify. When based on a search query, provide the Query parameter and optionally, the QueryContextId parameter. When based on item identifiers, provide a single appropriate parameter based on the identifier type, and specify the associated item value. MarketplaceId is always required. At least one of Query, SellerSKU, UPC, EAN, ISBN, JAN is also required. This operation returns a maximum of ten products and does not return non-buyable products. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 6 | 40 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.